requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.4.0
beautifulsoup4==4.12.3
certifi==2024.8.30
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
clip==1.0
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
exceptiongroup==1.2.2
fastapi==0.112.2
ffmpy==0.4.0
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
ftfy==6.2.3
googletrans-py==4.0.0
gradio==4.42.0
gradio_client==1.3.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hstspreload==2024.9.1
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.24.6
hyperframe==6.0.1
idna==2.10
importlib_resources==6.4.4
Jinja2==3.1.4
kiwisolver==1.4.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.2.1
numpy==2.1.1
open_clip_torch==2.26.1
orjson==3.10.7
packaging==24.1
pandas==2.2.2
pillow==10.4.0
protobuf==5.28.0
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
rfc3986==1.5.0
rich==13.8.0
ruff==0.6.3
safetensors==0.4.4
semantic-version==2.10.0
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.38.4
sympy==1.12
timm==1.0.9
tokenizers==0.19.1
tomlkit==0.12.0
torch==2.4.1+cu124
torchaudio==2.4.1+cu124
torchvision==0.19.1+cu124
tqdm==4.66.5
transformers==4.44.2
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
wcwidth==0.2.13
websockets==12.0
scikit-learn
scipy
fairscale
pycocoevalcap
google-generativeai